New: Add condition to set delay
Thank you for this wonderful tutorial. I used this snippet to do double tap in react native but it had some errors which I wanted to correct.
Considering the above code, when the user doesn't pass a prop of "delay" then the double tap doesn't work because it won't have the value of delay.
But I added that condition to check for the availability of delay so that whenever the user passes a custom value for delay it well set it to the passed value, else it will use that default one.

I would like also to convert this version into functional component, I'm I allowed to do it?